requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
Pillow>=10.1.0
django-environ>=0.11.2
django-crispy-forms>=2.1
crispy-tailwind>=1.0.0
django-filter>=23.5
django-tables2>=2.7.0
django-extensions>=3.2.3
python-dateutil>=2.8.2
openpyxl>=3.1.2
reportlab>=4.0.7
gunicorn>=21.2.0
whitenoise>=6.6.0